+++ CHANGELOG.md
@@ -4,5 +4,7 @@
 - MainAuto.py file
 - signal.py file
 
+### Added
+- added baseline code to signal.py that tunes to 315 MHz, recieves samples, and calculates their avg amplitude
 ### Added 
 - dsp_demodulation.py